backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
anthropic==0.18.1
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
langchain==0.1.6
langchain-anthropic==0.1.1
chromadb==0.4.22
sentence-transformers==2.3.1
numpy==1.26.3
pillow==10.2.0
pypdf2==3.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
